pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "asena-analysis"
version = "0.5.0"
description = "Defansif HTTP log analizi ve timeline (KVKK uyumlu maskeleme)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0.1",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "watchgod>=0.8",
  "anyio>=3.0",
  "requests>=2.31.0",
  "twilio>=8.10.0",
  "streamlit>=1.33.0",
  "plotly>=5.18.0",
  "streamlit-autorefresh>=1.0.1",
  "reportlab>=4.0.0",
]

[project.scripts]
asena = "asena:main"

[tool.setuptools]
py-modules = ["asena"]
